WIP: refactoring feature view dimensions.
